Added flashlight
This required adding the URP Asset, setting it to use the URP 2D renderer. From there you need to create materials for each object, set the material's shader to Universal Render Pipeline/2D/Sprite-Lit-Default, allowing the material to utilize the light rendering. After doing this for each object in the scene, and adding a spot light, you are able to modify the spot light into a cone and attach it to the player object